Adds AssemblyExtensions for assembly info
This is to replace the AssemblyInfo class as that class didn't allow for
the assembly to be specified. This reduced the use cases for retrieving
the assembly's information, such as Title or version info.
This update marks the AssemblyInfo class as Obsolete and will be removed
in a future update.