Updates AssemblyInfo & AssemblyExtensions
I've updated both classes to work properly with their respective TFM.
For example, Func was used in AssemblyInfo but is not implemented in
.NET 2.0. AssemblyExtensions is not compiled for .NET 2.0 as well since
extension methods are not available in that framework. I've added a
.NET 2.0 Console App to the test folder so I could test that everything
works (NUnit doesn't seem to allow targeting that framework).
I also added summary commits to the two classes.

Signed-off-by: Troy Geiger <troygeiger@users.noreply.github.com>